+++ context/app/static/js/components/detailPage/StatusIcon/ColoredStatusIcon/style.ts
@@ -1,3 +1,4 @@
+import React from 'react';
 import { styled } from '@mui/material/styles';
 import { InfoRounded, ErrorRounded, WarningRounded, CheckCircleRounded } from '@mui/icons-material';
 import { SvgIconProps } from '@mui/material/SvgIcon';