+++ context/app/static/js/components/workspaces/NewWorkspaceDialog/useCreateWorkspaceForm.ts
@@ -11,8 +11,8 @@ import {
   templatesField,
   workspaceJobTypeIdField,
 } from '../workspaceFormFields';
-import { useProtectedDatasetsForm, useTooManyDatasetsErrors } from '../formHooks';
-import { DEFAULT_JOB_TYPE } from '../constants';
+import { useProtectedDatasetsForm, useTooManyDatasetsErrors, useTooManyDatasetsWarnings } from '../formHooks';
+import { DEFAULT_JOB_TYPE, DEFAULT_TEMPLATE_KEY } from '../constants';
 
 export interface FormWithTemplates {
   templates: string[];
@@ -25,6 +25,7 @@ interface CreateWorkspaceFormTypes extends FormWithTemplates {
 
 interface UseCreateWorkspaceTypes {
   defaultName?: string;
+  defaultTemplate?: string;
 }
 
 const schema = z
@@ -32,7 +33,7 @@ const schema = z
   .partial()
   .required({ 'workspace-name': true, templates: true });
 
-function useCreateWorkspaceForm({ defaultName }: UseCreateWorkspaceTypes) {
+function useCreateWorkspaceForm({ defaultName, defaultTemplate }: UseCreateWorkspaceTypes) {
   const [dialogIsOpen, setDialogIsOpen] = useState(false);
   const createTemplateNotebooks = useTemplateNotebooks();
 
@@ -45,7 +46,7 @@ function useCreateWorkspaceForm({ defaultName }: UseCreateWorkspaceTypes) {
     defaultValues: {
       'workspace-name': defaultName ?? '',
       'protected-datasets': '',
-      templates: [],
+      templates: [defaultTemplate ?? DEFAULT_TEMPLATE_KEY],
       workspaceJobTypeId: DEFAULT_JOB_TYPE,
     },
     mode: 'onChange',
@@ -79,8 +80,14 @@ function useCreateWorkspaceForm({ defaultName }: UseCreateWorkspaceTypes) {
 function useCreateWorkspaceDatasets() {
   const { errorMessages: protectedDatasetsErrorMessages, selectedRows, ...rest } = useProtectedDatasetsForm();
   const tooManyDatasetsErrorMessages = useTooManyDatasetsErrors({ numWorkspaceDatasets: selectedRows.size });
+  const tooManyDatasetsWarningMessages = useTooManyDatasetsWarnings({ numWorkspaceDatasets: selectedRows.size });
 
-  return { errorMessages: [...protectedDatasetsErrorMessages, ...tooManyDatasetsErrorMessages], selectedRows, ...rest };
+  return {
+    errorMessages: [...protectedDatasetsErrorMessages, ...tooManyDatasetsErrorMessages],
+    warningMessages: [...tooManyDatasetsWarningMessages],
+    selectedRows,
+    ...rest,
+  };
 }
 
 export { useCreateWorkspaceForm, useCreateWorkspaceDatasets, type CreateWorkspaceFormTypes };