+++ context/app/static/js/components/workspaces/WorkspaceLaunchStopButtons/WorkspaceLaunchStopButtons.tsx
@@ -3,19 +3,20 @@ import Button, { ButtonProps } from '@mui/material/Button';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 
-import { useSnackbarActions } from 'js/shared-styles/snackbars';
 import { useWorkspacesList } from 'js/components/workspaces/hooks';
 import { isRunningWorkspace, findRunningWorkspace } from 'js/components/workspaces/utils';
 import { useLaunchWorkspaceStore } from 'js/stores/useWorkspaceModalStore';
+import { useSnackbarActions } from 'js/shared-styles/snackbars';
 import { Alert } from 'js/shared-styles/alerts';
+import { isWorkspaceAtDatasetLimit } from 'js/helpers/functions';
 import { MergedWorkspace } from '../types';
 
 interface WorkspaceButtonProps {
   workspace: MergedWorkspace;
   handleStopWorkspace: (workspaceId: number) => Promise<void>;
   isStoppingWorkspace: boolean;
-  disableLaunch?: boolean;
-  disableStop?: boolean;
+  showLaunch?: boolean;
+  showStop?: boolean;
   button: ElementType<ButtonProps>;
 }
 
@@ -24,12 +25,15 @@ function StopWorkspaceButton({
   handleStopWorkspace,
   button: ButtonComponent,
   isStoppingWorkspace,
-}: Omit<WorkspaceButtonProps, 'disableLaunch' | 'disableStop'>) {
+}: Omit<WorkspaceButtonProps, 'showLaunch' | 'showStop'>) {
   const { toastError } = useSnackbarActions();
+
   const currentWorkspaceIsRunning = isRunningWorkspace(workspace);
+
   if (!currentWorkspaceIsRunning) {
     return null;
   }
+
   return (
     <ButtonComponent
       type="button"
@@ -52,9 +56,11 @@ function StopWorkspaceAlertButton(props: ButtonProps) {
 
 function StopWorkspaceAlert() {
   const { handleStopWorkspace, isStoppingWorkspace, workspacesList } = useWorkspacesList();
-  const runningWorkspace = findRunningWorkspace(workspacesList);
 
-  if (!runningWorkspace) {
+  const runningWorkspace = findRunningWorkspace(workspacesList);
+  const runningWorkspaceAtMaxDatasets = runningWorkspace && isWorkspaceAtDatasetLimit(runningWorkspace);
+
+  if (!runningWorkspace || runningWorkspaceAtMaxDatasets) {
     return null;
   }
 
@@ -84,7 +90,7 @@ function StopWorkspaceAlert() {
 }
 
 function WorkspaceLaunchStopButtons(props: WorkspaceButtonProps) {
-  const { workspace, button: ButtonComponent, disableLaunch = false, disableStop = false } = props;
+  const { workspace, button: ButtonComponent, showLaunch = false, showStop = false } = props;
   const { open, setWorkspace } = useLaunchWorkspaceStore();
   if (workspace.status === 'deleting') {
     return (
@@ -95,10 +101,12 @@ function WorkspaceLaunchStopButtons(props: WorkspaceButtonProps) {
   }
   return (
     <Stack direction="row" spacing={2}>
-      {!disableStop && <StopWorkspaceButton {...props} />}
-      {!disableLaunch && (
+      {showStop && <StopWorkspaceButton {...props} />}
+      {showLaunch && (
         <Button
           type="button"
+          variant="contained"
+          color="primary"
           onClick={() => {
             setWorkspace(workspace);
             open();