+++ context/app/static/js/shared-styles/dropdowns/IconDropdownMenuButton/index.js
@@ -1,0 +1,3 @@
+import IconDropdownMenuButton from './IconDropdownMenuButton';
+
+export default IconDropdownMenuButton;